+++ gdal/port/cpl_vsil_stdin.cpp
@@ -131,7 +131,7 @@ int VSIStdinHandle::ReadAndCache( void* pBuffer, int nToRead )
 {
     CPLAssert(nCurOff == nRealPos);
 
-    int nRead = fread(pBuffer, 1, nToRead, stdin);
+    int nRead = static_cast<int>(fread(pBuffer, 1, nToRead, stdin));
 
     if (nRealPos < BUFFER_SIZE)
     {
@@ -160,7 +160,7 @@ int VSIStdinHandle::Seek( vsi_l_offset nOffset, int nWhence )
     if (nRealPos < BUFFER_SIZE )
     {
         nRealPos += fread(pabyBuffer + nRealPos, 1, BUFFER_SIZE - (int)nRealPos, stdin);
-        nBufferLen = nRealPos;
+        nBufferLen = static_cast<int>(nRealPos);
     }
 
     if (nWhence == SEEK_END)
@@ -366,7 +366,7 @@ int VSIStdinFilesystemHandler::Stat( const char * pszFilename,
     {
         VSIStdinInit();
         if (nBufferLen == 0)
-            nRealPos = nBufferLen = fread(pabyBuffer, 1, BUFFER_SIZE, stdin);
+            nRealPos = nBufferLen = static_cast<int>(fread(pabyBuffer, 1, BUFFER_SIZE, stdin));
 
         pStatBuf->st_size = nBufferLen;
     }
